Split everything into indiviual files
